two minor tweaks

+++ OndselSolver/FullVector.h
@@ -232,7 +232,7 @@ namespace MbD {
 	{
 		s << "FullVec{";
 		s << this->at(0);
-		for (size_t i = 1; i < (int)this->size(); i++)
+		for (size_t i = 1; i < (size_t)this->size(); i++)
 		{
 			s << ", " << this->at(i);
 		}
